skip unnecessary module compilations. SA 9.4 -> 1.0.12, bbfreeze->pyinstaller

+++ bqcore/bq/setup/module_setup.py
@@ -15,7 +15,7 @@ from bq.util.converters import asbool
 from mako.template import Template
 import bbfreeze
 
-BISQUE_DEPS = map (functools.partial(os.path.join, '../../external'), [ "bisque.jar", "jai_codec.jar", "jai_core.jar", "jai_imageio.jar", "clibwrapper_jiio.jar"])
+#BISQUE_DEPS = map (functools.partial(os.path.join, '../../external'), [ "bisque.jar", "jai_codec.jar", "jai_core.jar", "jai_imageio.jar", "clibwrapper_jiio.jar"])
 
 
 class SetupError(Exception):
@@ -25,6 +25,11 @@ def verbose_call (*args, **kw):
     if kw.pop('verbose',None):
         print args
     return check_call (*args, **kw)
+
+
+def needs_update (output, dependency):
+    return not os.path.exists (output) or os.path.getmtime (dependency) > os.path.getmtime (output)
+
 
 def ensure_matlab(params):
     'make sure matlab is enabled and in the path'
@@ -76,7 +81,7 @@ def mex_compile (command_list, where = None, **kw):
         os.chdir (cwd)
     return ret == 0
 
-def matlab (command, where = None, params={ }):
+def matlab (command, where = None, params=None):
     'run matlab with a command'
     if not require('runtime.matlab_home', params):
         return False
@@ -127,27 +132,22 @@ def matlab_setup(main_path, files = [], bisque_deps = False, dependency_dir = "m
     if main_path.endswith (".m"):
         main_path = main_path [0:-2]
 
-    #try:
-    #    script_name='matlab_launcher'
-    #    template = Template(filename=os.path.abspath (os.path.join('..','..', 'config','templates','matlab_launcher.tmpl')))
-    #    with open(script_name, 'wb') as f:
-    #        f.write(template.render())
-    #        os.chmod (script_name, 0744)
-    #except Exception,e:
-    #    log.exception ("while createing matlab_launcher")
-    #    print ("Could not create matlab launcher script %s" % e)
-
-
     main_name = os.path.basename(main_path)
     if bisque_deps:
         files.extend (BISQUE_DEPS)
     if len(files):
         copy_files (files, '.')
-    if not os.path.exists (dependency_dir): os.mkdir (dependency_dir)
+    if not os.path.exists (dependency_dir):
+        os.mkdir (dependency_dir)
 
-    if mcc(main_path + '.m', '-d', dependency_dir, '-m', '-C', '-R', '-nodisplay', '-R', '-nosplash', *rest):
-        main = main_name + ext_map.get(os.name, '')
-        ctf  = main_name + '.ctf'
+    source = main_path + '.m'
+    main = main_name + ext_map.get(os.name, '')
+    ctf  = main_name + '.ctf'
+    if not needs_update (ctf, source):
+        print "Skipping matlab compilation %s is newer than %s" % (ctf, source)
+        return 0
+
+    if  mcc(source, '-d', dependency_dir, '-m', '-C', '-R', '-nodisplay', '-R', '-nosplash', *rest):
         if os.path.exists (main):
             os.unlink (main)
         shutil.copyfile(os.path.join(dependency_dir, main), main)
@@ -159,7 +159,7 @@ def matlab_setup(main_path, files = [], bisque_deps = False, dependency_dir = "m
         return 0
     return 1
 
-def python_setup_bbfreeze(scripts,  package_scripts =True, dependency_dir = 'pydist', params = {} ):
+def python_setup_bbfreeze(scripts,  package_scripts =True, dependency_dir = 'pydist', params = None ):
     """compile python dependencies into a package
 
     if package_script is true then a runner scripts will be generated
@@ -174,7 +174,7 @@ def python_setup_bbfreeze(scripts,  package_scripts =True, dependency_dir = 'pyd
     fr()
     if not package_scripts:
         return
-    data = dict(params)
+    data = dict(params or {})
 
     for script in scripts:
         script_name = os.path.splitext(script)[0]
@@ -182,7 +182,8 @@ def python_setup_bbfreeze(scripts,  package_scripts =True, dependency_dir = 'pyd
         # THIS os.path.join needs to be replaced by pkg_resources or pkg_util
         # when the toplevel is packaged
         try:
-            template = Template(filename=os.path.abspath (os.path.join('..','..', 'config','templates','python_launcher.tmpl')))
+            filename=os.path.abspath(find_config_path('templates/python_launcher.tmpl'))
+            template = Template(filename=filename)
             with open(script_name, 'wb') as f:
                 f.write(template.render(script = data['script']))
                 os.chmod (script_name, 0744)
@@ -191,7 +192,7 @@ def python_setup_bbfreeze(scripts,  package_scripts =True, dependency_dir = 'pyd
             print ("Could not create python launcher script %s" % e)
 
 
-def python_setup(scripts,  package_scripts =True, dependency_dir = 'pydist', params = {} ):
+def python_setup(scripts,  package_scripts =True, dependency_dir = 'pydist', params = None ):
     """compile python dependencies into a package
 
     if package_script is true then a runner scripts will be generated
@@ -202,12 +203,15 @@ def python_setup(scripts,  package_scripts =True, dependency_dir = 'pydist', par
     #fr.include_py = False
     if not isinstance(scripts, list):
         scripts = [ scripts ]
+    if not any (needs_update (dependency_dir, x) for x in scripts):
+        print "Skippping python packaging step"
+        return
     for script in scripts:
         cmd.append (script)
     if os.path.exists (dependency_dir):
         shutil.rmtree (dependency_dir)
     check_call (cmd)
-    data = dict(params)
+    data = dict(params or {})
 
     for script in scripts:
         script_name = os.path.splitext(script)[0]
@@ -215,7 +219,8 @@ def python_setup(scripts,  package_scripts =True, dependency_dir = 'pydist', par
         # THIS os.path.join needs to be replaced by pkg_resources or pkg_util
         # when the toplevel is packaged
         try:
-            template = Template(filename=os.path.abspath (os.path.join('..','..', 'config','templates','python_launcher.tmpl')))
+            filename=os.path.abspath(find_config_path('templates/python_launcher.tmpl'))
+            template = Template(filename=filename)
             with open(script_name, 'wb') as f:
                 f.write(template.render(script = data['script']))
                 os.chmod (script_name, 0744)
@@ -227,6 +232,7 @@ def python_setup(scripts,  package_scripts =True, dependency_dir = 'pydist', par
 
 def ensure_binary(exe):
     'make sure executable is available to the module'
+    # pylint: disable=no-name-in-module, import-error, no-member
     import distutils.spawn
     p = distutils.spawn.find_executable(exe)
     if p is None:
@@ -240,6 +246,7 @@ def require(expression, params, throws = True):
     """
 
     valid = True
+    params = params or {}
     if not isinstance(expression,list):
         expression = [ expression ]
 
@@ -299,6 +306,7 @@ def docker_setup (image, command, base, params):
     hub = params.get ('docker.hub', None)
     if hub:
         hub_image = '%s/%s' % (hub, image)
+        print "Pushing docker %s to %s" % ( hub_image, hub)
         check_call(['docker', 'tag', '-f', image, hub_image])
         check_call(['docker', 'push', hub_image])
